requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
mypy==0.910
mypy-extensions==0.4.3
pyzmq==22.3.0
requests==2.26.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
